// front-end/crud-coppel/src/app/Model/Validacion.ts
export class Validacion {
    validSKU: boolean = true;
    validNombre: boolean = true;
    validMarca: boolean = true;
    validModelo: boolean = true;
    validDepartamento: boolean = true;
    validClase: boolean = true;
    validFamilia: boolean = true;
    validStock: boolean = true;
    validCantidad: boolean = true;
}